# application/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Simanja (Sistem Management Jamaah)",
        "description": "This is developped using Swagger UI.  You can find\r\nout more about Swagger at\r\n[http://swagger.io](http://swagger.io) or on\r\n[irc.freenode.net, #swagger](http://swagger.io/irc/).",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "email": "iqbalhanif313@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/setup/level": {
            "get": {
                "tags": [
                    "setup"
                ],
                "summary": "Show Setup Level information",
                "operationId": "setup/level",
                "responses": {
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "api_key": [
                            "write:user",
                            "read:user"
                        ]
                    }
                ]
            },
            "post": {
                "tags": [
                    "setup/level"
                ],
                "summary": "Store Setup Level",
                "operationId": "setup/level/store",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "description": "Store Setup Level",
                                        "type": "string",
                                        "example": "DRH"
                                    },
                                    "nama": {
                                        "description": "Store Setup Level",
                                        "type": "string",
                                        "example": "Daerah"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "api_key": [
                            "write:user",
                            "read:user"
                        ]
                    }
                ]
            }
        },
        "/api/setup/level/{id}": {
            "get": {
                "tags": [
                    "setup/level"
                ],
                "summary": "Show Detail Setup Level",
                "operationId": "setup/level/{id}",
                "responses": {
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "api_key": [
                            "write:user",
                            "read:user"
                        ]
                    }
                ]
            },
            "put": {
                "tags": [
                    "setup/level"
                ],
                "summary": "Update Setup Level",
                "operationId": "setup/level/update",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "description": "Update Setup Level",
                                        "type": "string",
                                        "example": "DRH"
                                    },
                                    "nama": {
                                        "description": "Update Setup Level",
                                        "type": "string",
                                        "example": "Daerah"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "api_key": [
                            "write:user",
                            "read:user"
                        ]
                    }
                ]
            },
            "delete": {
                "tags": [
                    "setup/level"
                ],
                "summary": "Delete Setup Level",
                "operationId": "setup/level/{id}/delete",
                "responses": {
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "api_key": [
                            "write:user",
                            "read:user"
                        ]
                    }
                ]
            }
        },
        "/api/setup/provinsi": {
            "get": {
                "tags": [
                    "setup/provinsi"
                ],
                "summary": "Show Setup Provinsi information",
                "operationId": "provinsi",
                "responses": {
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "api_key": [
                            "write:user",
                            "read:user"
                        ]
                    }
                ]
            },
            "post": {
                "tags": [
                    "setup/provinsi"
                ],
                "summary": "Store Setup Provinsi",
                "operationId": "setup/provinsi/store",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "description": "Store Setup Provinsi",
                                        "type": "string",
                                        "example": "DRH"
                                    },
                                    "nama": {
                                        "description": "Store Setup Provinsi",
                                        "type": "string",
                                        "example": "Daerah"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "api_key": [
                            "write:user",
                            "read:user"
                        ]
                    }
                ]
            }
        },
        "/api/setup/provinsi/{id}": {
            "get": {
                "tags": [
                    "setup/provinsi"
                ],
                "summary": "Show Detail Setup Provinsi",
                "operationId": "setup/provinsi/{id}",
                "responses": {
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "api_key": [
                            "write:user",
                            "read:user"
                        ]
                    }
                ]
            },
            "put": {
                "tags": [
                    "setup/provinsi"
                ],
                "summary": "Update Setup Provinsi",
                "operationId": "setup/provinsi/update",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "description": "Update Setup Provinsi",
                                        "type": "string",
                                        "example": "DRH"
                                    },
                                    "nama": {
                                        "description": "Update Setup Provinsi",
                                        "type": "string",
                                        "example": "Daerah"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "api_key": [
                            "write:user",
                            "read:user"
                        ]
                    }
                ]
            },
            "delete": {
                "tags": [
                    "setup/provinsi"
                ],
                "summary": "Delete Setup Provinsi",
                "operationId": "setup/provinsi/{id}/delete",
                "responses": {
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "api_key": [
                            "write:user",
                            "read:user"
                        ]
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "api_key": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}